perf(minifier): normalize `undefined` to `void 0` before everything else
so subsequent code don't need to lookup `undefined`.
